requirements.txt
Flask==1.0.2
python-socketio==5.3.0
requests==2.21.0
tornado==4.4